moved emitter to vergen-lib, docs updates
