+++ vergen-gitcl/src/lib.rs
@@ -45,9 +45,9 @@
 //!
 //! [build-dependencies]
 //! # All features enabled
-//! vergen-gitcl = { version = "1.0.0-beta.0", features = ["build", "cargo", "rustc", "si"] }
+//! vergen-gitcl = { version = "1.0.0", features = ["build", "cargo", "rustc", "si"] }
 //! # or
-//! vergen-gitcl = { version = "1.0.0-beta.0", features = ["build"] }
+//! vergen-gitcl = { version = "1.0.0", features = ["build"] }
 //! # if you wish to disable certain features
 //! ```
 //!
@@ -458,6 +458,14 @@ let build = BuildBuilder::default().build_timestamp(true).build()?;"##
 
 #[cfg(test)]
 use {lazy_static as _, regex as _, temp_env as _};
+// This is here to appease the `unused_crate_dependencies` lint
+#[cfg(not(any(
+    feature = "build",
+    feature = "cargo",
+    feature = "rustc",
+    feature = "si"
+)))]
+use vergen as _;
 
 mod gitcl;
 
@@ -471,7 +479,6 @@ pub use vergen::CargoBuilder;
 pub use vergen::CpuRefreshKind;
 #[cfg(feature = "cargo")]
 pub use vergen::DependencyKind;
-pub use vergen::Emitter;
 #[cfg(feature = "si")]
 pub use vergen::MemoryRefreshKind;
 #[cfg(feature = "si")]
@@ -486,3 +493,4 @@ pub use vergen_lib::AddCustomEntries;
 pub use vergen_lib::CargoRerunIfChanged;
 pub use vergen_lib::CargoWarning;
 pub use vergen_lib::DefaultConfig;
+pub use vergen_lib::Emitter;
