+++ vergen-gix/Cargo.toml
@@ -20,7 +20,7 @@ skip_optional_dependencies = true
 default = []
 build = ["vergen/build"]
 cargo = ["vergen/cargo"]
-emit_and_set = ["vergen/emit_and_set"]
+emit_and_set = ["vergen-lib/emit_and_set"]
 rustc = ["vergen/rustc"]
 unstable = ["vergen/unstable"]
 si = ["vergen/si"]
@@ -37,7 +37,7 @@ time = { version = "0.3.23", features = [
     "local-offset",
     "parsing",
 ] }
-vergen = { version = "9.0.0-beta.0", path = "../vergen", default-features = false }
+vergen = { version = "9.0.0-beta.2", path = "../vergen", default-features = false }
 vergen-lib = { version = "0.1.0", path = "../vergen-lib", features = ["git"] }
 
 [build-dependencies]
