+++ vergen-pretty/Cargo.toml
@@ -42,7 +42,7 @@ tracing = { version = "0.1.37", features = [
 [build-dependencies]
 anyhow = "1.0.70"
 rustversion = "1.0.12"
-vergen-gix = { version = "1.0.0-beta.1", features = [
+vergen-gix = { version = "1.0.0-beta.2", features = [
     "build",
     "cargo",
     "rustc",
