+++ vergen/Cargo.toml
@@ -19,17 +19,17 @@ skip_optional_dependencies = true
 [features]
 default = []
 build = ["time", "vergen-lib/build"]
-cargo = ["cargo_metadata", "regex", "vergen-lib/cargo"]
+cargo = ["cargo_metadata", "getset", "regex", "vergen-lib/cargo"]
 rustc = ["rustc_version", "vergen-lib/rustc"]
 si = ["sysinfo", "vergen-lib/si"]
 unstable = []
-emit_and_set = []
+emit_and_set = ["vergen-lib/emit_and_set"]
 
 [dependencies]
-anyhow = "1.0.72"
+anyhow = "1.0.79"
 cargo_metadata = { version = "0.18.1", optional = true }
 derive_builder = "0.13.0"
-getset = "0.1.2"
+getset = { version = "0.1.2", optional = true }
 regex = { version = "1.9.1", optional = true }
 rustc_version = { version = "0.4.0", optional = true }
 sysinfo = { version = "0.30.4", optional = true, default-features = false }
